requirements.txt
poetry==1.7.1
playwright==1.41.0
python-dotenv==1.0.0
tweepy==4.14.0
requests==2.31.0 